requirements.txt
asgiref==3.2.3
Django==3.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
